requirements.txt
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.12.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
distro==1.9.0
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
jiter==0.13.0
lxml==5.3.0
openai==1.54.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.33.1
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.8.3
pytest==8.3.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
